requirements.txt

reflex==0.8.24.post1
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.30.0
google-genai>=0.3.0
# Node.js is required for Reflex frontend tooling; install separately (see README).
